Update app to SDK 2.0
The old way to select the correct app no longer works as the appinfo.json file
now control the UUID, name and other stuff we want to change according to the
build-choices.

Therefore we introduce the configure.sh script which will create an
appinfo.json and src/config.h file matching the selected choices.

Apart from this, the only changes are those that were necessary to get the app
to work with SDK 2.0BETA4.